Allow signs/bamboo/cactus relying on support blocks being built to be pasted
canSurvive() could fail if the supporting render block had not been placed
yet. We can attempt to salvage this check by simulating it in the build's
eventual form, but it's not so straightforward.

Ultimately, the user should make the call and decide if placing the build in
the chosen area is logical